CaseContext signing out aware - hides any previous data
